# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Mô tả API",
        "contact": {
            "email": "email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng ký tài khoản",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fullName",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "a@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0912345678"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Hanoi"
                                    },
                                    "fullName": {
                                        "type": "string",
                                        "example": "Nguyen Van A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng ký thành công"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
        "/api/verify-email": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Xác thực email (kích hoạt tài khoản)",
                "operationId": "30ad750305b3a9051ad8902b774acaa0",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Mã token xác thực được gửi qua email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Abc123xyz456token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tài khoản đã được kích hoạt thành công.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tài khoản đã được kích hoạt thành công."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token không hợp lệ hoặc đã hết hạn.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token không hợp lệ hoặc đã hết hạn."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Đăng nhập người dùng",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Không có quyền"
                    }
                }
            }
        },
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Gửi link reset mật khẩu",
                "operationId": "254ceded43f5376d0fa782a33f5b7d14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "a@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link reset mật khẩu đã được gửi"
                    },
                    "500": {
                        "description": "Không thể gửi email"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset mật khẩu",
                "operationId": "cfc7166a8a49cec932e10496c2ecf013",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "a@gmail.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "abc123token"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset mật khẩu thành công"
                    },
                    "400": {
                        "description": "Token không hợp lệ hoặc đã hết hạn"
                    }
                }
            }
        },
        "/api/example": {
            "get": {
                "summary": "Lấy danh sách example",
                "operationId": "8dcd33e8185b7ca83e05be042cb88d93",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "List of all users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "username": {
                                                "type": "string",
                                                "example": "user1"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "user1@example.com"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "123456789"
                                            },
                                            "fullName": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "123 Main St"
                                            },
                                            "avatar": {
                                                "type": "string",
                                                "example": "avatars/user1.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/update-profile": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Cập nhật hồ sơ người dùng",
                "operationId": "ffb14df5b0041f66fb511391f845eb51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "phone",
                                    "fullName",
                                    "address"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "user1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "fullName": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật hồ sơ thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật hồ sơ thành công"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "user1"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "user1@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "123456789"
                                                },
                                                "fullName": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "123 Main St"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "avatars/user1.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/avatar": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Cập nhật ảnh đại diện người dùng",
                "operationId": "b6af0338f700c6913333f691a710e1d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật ảnh đại diện thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cập nhật ảnh đại diện thành công"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "avatars/123456.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Lấy thông tin người dùng theo ID",
                "operationId": "06175584a3eb3b8f3a4b51244a43c517",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Không tìm thấy người dùng"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Xóa người dùng",
                "operationId": "f8d7aea516ee464d0afee90f3a1a1722",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa người dùng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa người dùng thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Người dùng không tồn tại"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "User",
            "description": "User"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "description": "Nhập token theo định dạng: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}